feat: 그룹러닝 유저리스트 불러오기 기능추가

+++ TrackUs/TrackUs/Sources/Commons/Extensions/Array+CLLocationCoordinate2D.swift
@@ -21,7 +21,7 @@ extension Array where Element == CLLocationCoordinate2D {
         return CLLocationCoordinate2D(latitude: averageLatitude, longitude: averageLongitude)
     }
     
-    func getTotalDistanceBetweenPoints() -> Int {
+    func getTotalDistanceBetweenPoints() -> Double {
         return 0
     }
 }
